implement shortcodes search/replace in generate scripts
