+++ generator/scripts/download-preview/generate-download-preview.js
@@ -1,4 +1,6 @@
-export default async function generateDownloadPreview(signed=false, forDownload=false) {
+import { doShortcodesFromLocalstorage } from "../utils.js";
+
+export default async function generateDownloadPreview(signed = false, forDownload = false) {
 
     let signedFileName = "contract-signed.html"
 
@@ -19,7 +21,9 @@ export default async function generateDownloadPreview(signed=false, forDownload=
 
     // get main from editor (don't escape single quotes)
     let main = document.querySelector(".editor-container .ql-editor").innerHTML
-    // //
+    
+    // replace each shortcode w/ its value
+    main = doShortcodesFromLocalstorage(main)
 
     // js for unsigned contract
     let contract_script_unsigned = `
