+++ generator/scripts/utils.js
@@ -39,3 +39,31 @@ export const select = (...selectors) => selectors.map(selector => document.query
 // selectAll() usage: const [sel1, sel2] = selectAll(".class1", ".class2")
 // wrapper around querySelectorAll
 export const selectAll = (...selectors) => selectors.map(selector => document.querySelectorAll(selector));
+
+// Shortcodes
+export function doShortcodes(shortcodes, contentString) {
+    // Replace names with values in the provided string
+    shortcodes.forEach(item => {
+        const namePattern = new RegExp(escapeRegExp(item.name), 'g');
+        contentString = contentString.replace(namePattern, item.value);
+    });
+
+    return contentString;
+}
+export function doShortcodesFromLocalstorage(main) {
+    // localStorage.setItem('repeaterData', jsonString);
+    // localStorage.removeItem('repeaterData');
+    const shortcodesJsonString = localStorage.getItem('repeaterData');
+
+    if (shortcodesJsonString) {
+        const shortcodes = JSON.parse(shortcodesJsonString);
+        if (shortcodes[0]) { // console.log(shortcodes[0].name);
+            return doShortcodes(shortcodes, main)
+        }
+    }
+    return main;
+}
+// Function to escape special characters in a string for regex
+export function escapeRegExp(string) {
+    return string.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
+}
